requirements.txt
blinker==1.6.2
boto3==1.26.129
botocore==1.29.129
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
psycopg2==2.9.6
pyodbc==4.0.39
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.2
python-dotenv==0.21.0
